requirements.txt
# Core
pandas
polars
pyarrow
fastexcel
pydantic
SQLAlchemy
psycopg2-binary

# Jupyter / notebooks
notebook
ipykernel
jupyterlab

# Utilities
python-dotenv  # for reading .env